Deprecate move legacy normlizer classes - functionality better met with SMIRKS (via AMBIT atm)
