+++ legacy/pom.xml
@@ -104,7 +104,6 @@
             <groupId>org.openscience.cdk</groupId>
             <artifactId>cdk-smiles</artifactId>
             <version>${project.version}</version>
-            <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>org.openscience.cdk</groupId>
