+++ legacy/src/main/java/org/openscience/cdk/normalize/CDKUtilities.java
@@ -2,23 +2,24 @@
  *
  * Contact: cdk-devel@lists.sourceforge.net
  *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public License
- * as published by the Free Software Foundation; either version 2.1
- * of the License, or (at your option) any later version.
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation; either version 2.1 of the License, or (at
+ * your option) any later version. All we ask is that proper credit is given
+ * for our work, which includes - but is not limited to - adding the above
+ * copyright notice to the beginning of your source code files, and to any
+ * copyright notice that you may distribute with programs based on this work.
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public
+ * License for more details.
  *
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 U
  */
-package org.openscience.cdk.tools;
-
-import java.util.List;
+package org.openscience.cdk.normalize;
 
 import org.openscience.cdk.CDKConstants;
 import org.openscience.cdk.aromaticity.Aromaticity;
@@ -31,6 +32,8 @@ import org.openscience.cdk.ringsearch.AllRingsFinder;
 import org.openscience.cdk.tools.manipulator.AtomContainerManipulator;
 import org.openscience.cdk.tools.manipulator.RingSetManipulator;
 
+import java.util.List;
+
 /**
  * Utility class written by Todd Martin, for help in his QSAR descriptors and SMILES
  * parser. Seems to have overlap with, at least, cdk.normalize.Normalizer.
@@ -41,8 +44,10 @@ import org.openscience.cdk.tools.manipulator.RingSetManipulator;
  * @cdk.module extra
  * @cdk.githash
  *
- * @see        org.openscience.cdk.normalize.Normalizer
+ * @see        Normalizer
+ * @deprecated better accomplished with SMARTS patterns or simple REGEX
  */
+@Deprecated
 public class CDKUtilities {
 
     public static String fixSmiles(String Smiles) {
