+++ legacy/src/main/java/org/openscience/cdk/normalize/Normalizer.java
@@ -1,32 +1,27 @@
 /*  
- *  Copyright (C) 2004-2010  The Chemistry Development Kit (CDK) project
- *                     2014  Mark B Vine (orcid:0000-0002-7794-0426)
+ * Copyright (C) 2004-2010  The Chemistry Development Kit (CDK) project
+ *                    2014  Mark B Vine (orcid:0000-0002-7794-0426)
  *
- *  Contact: cdk-devel@lists.sourceforge.net
+ * Contact: cdk-devel@lists.sourceforge.net
  *
- *  This program is free software; you can redistribute it and/or
- *  modify it under the terms of the GNU Lesser General Public License
- *  as published by the Free Software Foundation; either version 2.1
- *  of the License, or (at your option) any later version.
- *  All we ask is that proper credit is given for our work, which includes
- *  - but is not limited to - adding the above copyright notice to the beginning
- *  of your source code files, and to any copyright notice that you may distribute
- *  with programs based on this work.
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation; either version 2.1 of the License, or (at
+ * your option) any later version. All we ask is that proper credit is given
+ * for our work, which includes - but is not limited to - adding the above
+ * copyright notice to the beginning of your source code files, and to any
+ * copyright notice that you may distribute with programs based on this work.
  *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU Lesser General Public License for more details.
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public
+ * License for more details.
  *
- *  You should have received a copy of the GNU Lesser General Public License
- *  along with this program; if not, write to the Free Software
- *  Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
- *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 U
  */
 package org.openscience.cdk.normalize;
-
-import java.util.Iterator;
-import java.util.List;
 
 import org.openscience.cdk.exception.CDKException;
 import org.openscience.cdk.exception.InvalidSmilesException;
@@ -40,6 +35,9 @@ import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.NodeList;
 
+import java.util.Iterator;
+import java.util.List;
+
 /**
  * Adjusts parts of an AtomContainer to the configuration of a fragment.
  *
@@ -47,7 +45,9 @@ import org.w3c.dom.NodeList;
  * @cdk.created   2004-03-04
  * @cdk.module    smiles
  * @cdk.githash
+ * @deprecated The functionality provided by with class is better suited to SMIRKS
  */
+@Deprecated
 public class Normalizer {
 
     /**
