+++ legacy/src/main/resources/org/openscience/cdk/normalize/fixsmiles.xml
